+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/dmp_network_service/wallet_store/interfaces/WalletStoreManager.java
@@ -88,7 +88,7 @@ public interface WalletStoreManager {
                               List<byte[]> previewImageList,
                               boolean hasVideoPreview,
                               List<URL> videoPreviews,
-                              int skinSizeInBytes,
+                              long skinSizeInBytes,
                               com.bitdubai.fermat_api.layer.dmp_identity.designer.interfaces.Designer designer,
                               boolean isDefault);
